Working on fleshing out finer details in Goat.
Implemented a global thread counting object to deal with potential multiple threads
when the program exits. The idea is that when main thread calls invoke new threads
they also add threads to the counter, and when those threads finish they should
remove threads from the counter. Then, on exit simply need to check the value of
the the thread counter and issue a warning if one or more threads are running.

Also need to implement a _changed flag on all top-level window objects that allow
modification of one or more aspects, and to issue a warning when the user quits
after making changes (i.e. presses 'cancel' instead of 'submit').

Lastly, would like to be able to prevent some windows from being opened twice,
which could either be done through instance variables on the main root object or,
more likely, through yet another global. It might be useful to create some sort of
hierarchy so that widgets that could potentially change other widgets can give
them the chance to update themselves when changes are made to databases they care
about - perhaps a general class containing objects with more specific properties
and ways to invoke general functions?