Implemented text label clipping on redraw.
Tested out clipping long label text to fit inside a minimum length window using
the database viewer interface. Key is to bind the <Configure> event to the display
function and call the widgets winfo_width() methods, convert between returned
value in pixels and approximate width of the text (using the '0' character for a
standard reference), and then clipping any lines that are longer than the
available width.

This appears to work, and could be easily implemented for other such windows.
Wondering if eventually all such viewer widgets could inherit from a common
superclass? It would still be the job of individual subclasses to determine how to
concatenate their object's info, but the functions involved in actually displaying
it would appear to be common to all.