Working on implementing analyses.
Starting with analyses now. Reciprocal BLAST appears to work, but the implementation
is a bit hacky. Because both the forward and reverse search are threaded, need to
add a "continue" callback to the forward search and repress closing the parent
window in order to actually get the reverse search to work properly and allow time
for the callback in each case to parse the output. This could be improved by either
altering the implementation of searches in general or by changing the non-threaded
version of search runner to also work with a new window and progress bars. This way,
could simply call this non-threaded version from within the reciprocal BLAST thread
and the linear execution would ensure parsing happened before reverse BLAST.

Also, need to go back and re-investigate the implementation of the status bar for
searches. Set it to 1 initially because the window pops up before the first search
finishes, but it does occassionally report 1 search over the actual number (e.g.
'search 3 of 2').

At this point, should probably commit all changes and then go back and try to even
out the implementation for all segments to use the same mechanics (e.g. label clip
on resize, unique ids and recursion in treeview, etc.). After this, should deal with
searches a little bit better, and also figure out how to work with multiple query
types. Then, need to finish all remaining options without code and begin testing.