Re-working queries.
Changing implementation of queries in Goat. Splitting up sets and queries in terms
of interface and also databases. Also, adding specific interfaces for dealing with
different types of queries (especially seqs and hmms).

There are some challenging aspects to this, namely that to properly implement HMM
queries currently nesting GUI several windows deep, which might annoy user. Will
leave as-is for now, but think of better ways of dealing with this in the future.

Still need to work out:
-interfaces for modifying hmm-based queries
-dealing with searches for submitting raccs
    -if not adding raccs, add queries directly to widget
    -else, run searches in separate thread and add to widget with callback