Changed searches and adding redundant accessions for queries.
Making slight but important changes for how searches are run. Trying to separate
threading from the actual execution of searches and also make it easier to run
multiple searches in the same window. Idea is to have the progress frame in all
cases, and to run sequential searches within the thread (if threaded) to prevent
having to callback to start subsequent searches. Now, searches run and parse
sequentially, and callbacks are run only after all searches finish.

Search runner now uses per-query alphabet and db-type information (the latter if
specified), as queries for racc searches may not all be of the same type. For normal
searches this won't make a difference.

Finally, need to test this by adding some queries. Assuming it works, can move
on to interfaces for modifying queries again.