Continuing work on queries.
Fleshed out HMM queries a little bit more so that code reads associated sequence/
msa files and determines the number of sequences. Still need to make it so that
user is warned if number of sequences in both files is different but it is a start.

HMM queries appear to be properly added and their attribute displayed, but now
need to work on code to modify their attributes (and most importantly to deal with
modifying redundant accessions of their associated queries - picturing the same
interface as with sequence-based queries, but changing the listbox entries each
time the combobox selection changes).