HMM query modifications started.
Started basic outline of modifying HMM queries. Linked combobox to the listboxes
for specifying raccs for the associated queries, but will likely need to add a
function to clear the listboxes first before adding new entries. Also need to find
a way to store all of the different raccs for associated queries before hitting
'Submit' - cannot simply use listbox persistence, so maybe a dictionary of lists
would be more appropriate?