Dealing with searches again now.
Making improvements/changes to searches to fit in better with recent changes. First
step is to add support for both query and DB sets. Made it so that query windows
change their display when different search algorithms are chosen. Idea here is to
ensure that only appropriate queries are used for each search.

Next step is to add support for record sets and then move on to running searches
with the new threaded progress frame code and changing intermediate query code so
that these queries are stored in a separate database.