Working now on outputting sequences from previous results.
Finished initial implementation of getting sequences from summaries. Like other
code in Goat, attempted to separate GUI and non-GUI code, so that the latter can
be called from other parts of the interface (e.g. following a search so that the
positive hits can be gathered and used to build a HMM).

Will work next on additional analyses.