Containuing work on analyses.
Finished initial code for HMMer forward search/BLAST reverse search analysis. Only
real issue that arose is how to deal with choosing the appropriate record for
reverse BLAST searches. Current implementation is to choose one global record, and
default to the first associated_query record for each HMM query if this is not
specified. Would be nice to have finer control over each HMM object instead, but
this would require changing boilerplate code for query chooser or adding additional
windows/panels.

Next step is to implement a full BLAST/HMMer analysis. Here, believe the appropriate
course of action is to use the original querie's record as the chosen record for
rBLAST following HMMer.

Another question to sort out is how to summarize when there are multiple searches?
Likely just leave it up to the user to choose the fwd and reverse searches, but in
terms of redundant accs, this would have to default to the associated_querie's
raccs in case of HMMs and this might be blank? Also, what if the chosen record is
not present in the HMMQuery object's records (for assoc_queries?). Would be better
then to choose an appropriate record...