Minor addition to sequences from summary code.
Made an option to add the relevant query sequence as well. Should be useful for
downstream analyses and is almost a necessity for code dealing with creating
intermediate MSA/HMM files for searches. Could add code in each of these places
but much easier to encapsulate everything in the single class.