Working on full BLAST/HMMer analysis.
Adding code to run a full forward/reverse BLAST followed by a full forward HMMer/
reverse BLAST search. Main question to solve moving forward is how to associate
the forward query in the original search with the HMMer/rev BLAST search. Could
simply use the fact that the query name is present in the intermediate sequence
file? This might not be the best, but it is the simplest... alternatively, could
try to sort out something using a class-based system per Query but this would
be complicated with the current implementation of 'seqs_from_summary', which
takes whole summaries as input and not individual queries.