Full BLAST and HMMer analysis now works.
Finished and performed initial tests for code to run a full BLAST/HMMer analysis
consisting of forward and reverse BLAST, creation of intermediate HMMer file(s),
HMMer search into same group of databases as original search, and then finally a
reverse BLAST of those new queries.

In order to achieve this, changed the model of how HMMer searches work in terms
of reverse BLAST and redundant accessions for queries. All query objects have
a "spec_qid" and "spec_record" attribute, that can be set and changed, and then
is encapsulated in the result object at search time to capture the value that
was used on a per-search basis. This then forms the basis for performing reverse
searches and summarizing these. Still need to test summarizing the output of the
second search, but at least the searches are performed.

Next steps:
1) Test and make changes, if necessary, to summary code for the second search in
the above analysis
2) Work on code to sumamrize two summaries with the same query/db set
    -This will be critical moving forward to fully leverage power of the above
    analysis and others that produce multiple searches
3) Clean up all code
4) Consider changing the notion of name/ID for objects within GOAT?
    -Could use a single numeric counter to assign unique IDs to all objects
    regardless of their actual contents
    -How would this affect intermediate steps/searching through DBs/formatting
    output?
5) Test code with old PLoS One study to ensure searches run correctly
6) Try with a much larger dataset!