Summary code also works for HMMer searches.
Had to make some minor changes; first of all, had to add a 'title' attribute to
HMMer record and change 'evalue' to 'e' to be consistent with BLAST as layed out
in BioPython. Second, had to cast all numeric values within the HMMer record
object as either int or float. Lastly, and with potential unforeseen consequences,
made the check between fwd_qobj.identity and rev_qobj.original_query less
stringent, because the identity of the fwd hmmer object is not always going to
be the same, especially when performing the full blast/hmmer analysis. This should
be ok though, because the second check for rev_qobj.identity and fwd_uobj.int_queries
should still act as a failsafe.

Additional issues may become apparent with time. Next need to try and summarize
summaries.