Fixed previously unappreciated issue with linked scrollboxes.
For choosing redundant accessions (and other applications), wanted to link scroll
boxes in a way that the internal order of items was maintained when adding/removing
from either. Original implementation worked in the most obvious case, but was not
robust when adding/removing out of order. Improved by adding internal dicts not
only for item values but also for index (also, is the value dict still needed?).

This works in cases where only a starting set of items are added. In cases where
more items are added after instantiation, will have to come up with a way to add
these with appropriate indices.