Finished initial implementation of scrollsaw.
Currently only works for individual summaries and can only choose to separate based
on supergroup, but this will change moving forward. Next best idea would be to run
off of input files. Reason for this is because summarizing summaries cannot
currently handle counting different starting queries as "the same thing" for the
purposes of summarizing - Could probably change this, but could also simply get
the sequence files from each, get the union using external code, and then run
scrollsaw on the union files.

Another improvement that needs to happend is to add scrollsaw as an analysis
under the threaded search progress bar model. For some things (e.g. getting seqs,
the time is not prohibitive, but for scrollsaw it could take quite a while with
larger datasets and so would be useful to put in a thread).

Next step is to test on previous PLoS One dataset.