Working on updating settings implementation to work with ZODB DB.
Instead of having a separate pickled settings file, adding settings to
the same DB as everything else, along with a new GUI implementation to
view and modify the values of settings.

Basic implementation worked out. Each setting is a simple object stored
in a database and grouped into sets based on what the setting relates
to. GUI implementation is proving difficult however. Want to display
settings in justified rows for viewing and then give an option to change
settings which uses a entry rows or some similar thing.

Settings display as is, but having trouble lining up text as the width
of each character is different. Considered using a single width for the
"average" character, but may have to get more in depth with the actual
widths by measuring every single character. As this is an expensive
operation, can maybe work out a hash table or something?

Once display is done, can create a simple interface for modifying.